fix (api) : spoof detection service

+++ spoofing_detection_api/app/services/spoof_detection/checkers/glass.py
@@ -34,8 +34,8 @@ class GlassChecker:
     def preprocess(self, image: Image.Image) -> np.ndarray:
         img = image.resize((256, 256))
         img = np.array(img).astype(np.float32) / 255.0
-        mean = np.array([0.485, 0.456, 0.406])
-        std = np.array([0.229, 0.224, 0.225])
+        mean = np.array([0.485, 0.456, 0.406], dtype=np.float32)
+        std = np.array([0.229, 0.224, 0.225], dtype=np.float32)
         img = (img - mean) / std
         img = np.transpose(img, (2, 0, 1))
         img = np.expand_dims(img, axis=0)
